공지사항 게시판 상세보기(info)에 등록한 첨부파일 불러오기 구현 완료

+++ src/main/java/com/spring/mapper/NoticeAttachMapper.java
@@ -10,5 +10,5 @@ public interface NoticeAttachMapper {
 	
 	public void delete(String uuid);
 	
-	public List<NoticeAttachVO> findByBno(Long bno);
+	public List<NoticeAttachVO> findByBno(int bno);
 }
